Adds GitHub icon landing page
