+++ src/css/GoldifyLandingPage.css
@@ -32,6 +32,35 @@
   vertical-align: middle;
 }
 
+.follow-github-container {
+  margin: 0 auto 5% auto;
+  width: fit-content;
+  background-color: #212121;
+  border: 4px solid white;
+  box-shadow: 0 5px 10px 0 rgba(0, 0, 0, 0.4);
+  border-radius: 22px;
+}
+
+.follow-github-container img {
+  display: inline-block;
+  vertical-align: middle;
+  margin-right: 8px;
+  height: 32px;
+}
+
+.follow-github-message {
+  display: inline-block;
+  vertical-align: middle;
+  font-size: 18px;
+  font-weight: bolder;
+  color: white;
+}
+
+.follow-github-inner {
+  margin: 6px;
+  cursor: pointer;
+}
+
 .landing-goldify-logo {
   height: 100px;
   margin: 15px;
