+++ src/js/utils/constants.js
@@ -8,3 +8,4 @@ export const mediumTermTracksRecommended = 2;
 export const longTermTracksRecommended = 2;
 export const spotifyWebPlayerDomain = "https://open.spotify.com";
 export const spotifyHomePageUrl = "https://www.spotify.com/";
+export const goldifyGitHubUrl = "https://github.com/Goldify/Goldify";
